Boring code for creating the Board structure
The board constructor should return always the same board, with the only difference of number of infections cards on the player deck. It means that shuffling, collocation of players in Atlanta and original virus distribution will be done out of the constructor.

On this commit it is created the EndOfGameException, which will be thrown when there is a reason to finalize the game, with a parameter with the text.

All this text would be moved to a literal file on future commits.

Board test is empty, because is boring to do it now, as though too much on the Object structure until reach the final decision. Which is the decision? Board constructor will return a deterministic state on the constructor call. Logic for accessing to Board private fields will be done by special function, not by pure getters and setters, in order to encapsulate data on a logical way.

On BoardCreationHelper it will be implemented the functions to be access from the game flow and modify the state of the Board.