ConfigurePlayersOnBoard
On this method we are going to configure the players on the board and provide them with the initial set of cards.

It has been found a problem on the classes model and the way the Board is initialized.

On the actual flow, the constructor contains a parameter for the number of epidemic cards, and also, shuffles the cards. Because of the shuffling, it makes the output indeterministic, and because of the epidemic cards, it make more complex to draw only city cards.

About the modeling, we know that a Player will have only CityCards on their hand but on the player deck there are Epidemic and CityCards, then, the list is a BaseCard one, to gather both types, but make the draw much complex.

Class modeling looks good, but then, we will need to commit this changes on future commits to simplify the pattern in:

* Constructor on Board will be out of parameters
* The player deck will be not shuffle and without EPIDEMIC cards, which will be added after the Player configuration, as is done on analogic game, and as already happen with the ACTION_CARDS.
* As player will have ACTION_CARDS on the future, make sense to have a list of BaseCards